centralisation of the code for components
