+++ Grasshopper_UI/CallerComponent/IGH_VariableParameterComponent.cs
@@ -39,6 +39,7 @@ using System.Collections;
 using BH.Adapter;
 using BH.oM.Reflection.Debugging;
 using BH.UI.Base;
+using Grasshopper.Kernel.Parameters.Hints;
 
 namespace BH.UI.Grasshopper.Templates
 {
@@ -50,33 +51,85 @@ namespace BH.UI.Grasshopper.Templates
 
         public virtual bool CanInsertParameter(GH_ParameterSide side, int index)
         {
-            return false;
+            if (side == GH_ParameterSide.Input)
+                return Caller.CanAddInput();
+            else
+                return Caller.CanAddOutput();
         }
 
         /*************************************/
 
         public virtual bool CanRemoveParameter(GH_ParameterSide side, int index)
         {
-            return false;
+            if (side == GH_ParameterSide.Input)
+            {
+                if (Params.Input.Count <= index)
+                    return false;
+                else
+                    return Caller.CanRemoveInput(Params.Input[index].NickName);
+            }
+            else
+            {
+                if (Params.Output.Count <= index)
+                    return false;
+                else
+                    return Caller.CanRemoveOutput(Params.Output[index].NickName);
+            }  
         }
 
         /*************************************/
 
         public virtual IGH_Param CreateParameter(GH_ParameterSide side, int index)
         {
-            return new Param_GenericObject();
+            string name = GH_ComponentParamServer.InventUniqueNickname("xyzuvw", this.Params.Input);
+            Param_ScriptVariable param = new Param_ScriptVariable
+            {
+                NickName = name,
+                Name = name,
+                TypeHint = new GH_NullHint()
+            };
+
+            // Updating the caller with the parameter that Grasshopper just added
+            Caller.AddInput(index, param.NickName, param.Type());
+            return param;
         }
 
-        /*************************************/
+        /*******************************************/
 
         public virtual bool DestroyParameter(GH_ParameterSide side, int index)
         {
-            return true;
+            if (side == GH_ParameterSide.Input)
+            {
+                if (Params.Input.Count <= index)
+                    return true;
+                else
+                    Caller.RemoveInput(Params.Input[index].NickName);
+            }
+            else
+            {
+                if (Params.Output.Count <= index)
+                    return true;
+                else
+                    Caller.RemoveOutput(Params.Output[index].NickName);
+            }
+                return true;
         }
 
-        /*************************************/
+        /*******************************************/
 
-        public virtual void VariableParameterMaintenance() { }
+        public virtual void VariableParameterMaintenance()
+        {
+            foreach (IGH_Param param in Params.Input)
+            {
+                Param_ScriptVariable paramScript = param as Param_ScriptVariable;
+                if (paramScript != null)
+                {
+                    paramScript.ShowHints = true;
+                    paramScript.Hints = Engine.Grasshopper.Query.AvailableHints;
+                    paramScript.AllowTreeAccess = true;
+                }
+            }
+        }
 
         /*******************************************/
     }
