+++ Grasshopper_UI/Components/oM/CreateCustom.cs
@@ -40,82 +40,6 @@ namespace BH.UI.Grasshopper.Components
 
 
         /*******************************************/
-        /**** Constructors                      ****/
-        /*******************************************/
-
-        public CreateCustomComponent() : base()
-        {
-            this.Params.ParameterChanged += OnGHParamChanged;
-        }
-
-
-        /*******************************************/
-        /**** Override Methods                  ****/
-        /*******************************************/
-
-        public override bool CanInsertParameter(GH_ParameterSide side, int index)
-        {
-            return side == GH_ParameterSide.Input;
-        }
-
-        /*******************************************/
-
-        public override bool CanRemoveParameter(GH_ParameterSide side, int index)
-        {
-            return side == GH_ParameterSide.Input;
-        }
-
-        /*******************************************/
-
-        public override IGH_Param CreateParameter(GH_ParameterSide side, int index)
-        {
-            Param_ScriptVariable param = new Param_ScriptVariable
-            {
-                NickName = GH_ComponentParamServer.InventUniqueNickname("xyzuvw", this.Params.Input),
-                TypeHint = new GH_NullHint()
-            };
-
-            // Updating the caller with the parameter that Grasshopper just added
-            CreateCustomCaller caller = Caller as CreateCustomCaller;
-            if (caller != null && param != null)
-                caller.AddInput(index, param.NickName, param.Type());
-            return param;
-        }
-
-        /*******************************************/
-
-        public override bool DestroyParameter(GH_ParameterSide side, int index)
-        {
-            if (side == GH_ParameterSide.Output)
-                return true;
-
-            if (Params.Input.Count <= index)
-                return true;
-
-            // Updating the caller with the parameter that Grasshopper just removed
-            CreateCustomCaller caller = Caller as CreateCustomCaller;
-            if (caller != null)
-                caller.RemoveInput(Params.Input[index].NickName);
-            return true;
-        }
-
-        /*******************************************/
-
-        public override void VariableParameterMaintenance()
-        {
-            foreach (IGH_Param param in Params.Input)
-            {
-                Param_ScriptVariable paramScript = param as Param_ScriptVariable;
-                if (paramScript != null)
-                {
-                    paramScript.ShowHints = true;
-                    paramScript.Hints = Engine.Grasshopper.Query.AvailableHints;
-                    paramScript.AllowTreeAccess = true;
-                }
-            }
-        }
-
-        /*******************************************/
     }
 }
